requirements.txt
pydantic>=1.10.0,<2.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
boto3>=1.26.0
pandas>=1.5.3,<2.0.0
pyarrow>=11.0.0,<12.0.0
s3fs>=2023.11.1,<2024.0.0
python-dotenv